Ported SaveScreenshotCommand
It now uses the new command.execute(...) syntax
