+++ automation/Commands/browser_commands.py
@@ -227,16 +227,30 @@ class BrowseCommand(BaseCommand):
                 pass
 
 
-def save_screenshot(visit_id, browser_id, driver, manager_params, suffix=""):
-    """ Save a screenshot of the current viewport"""
-    if suffix != "":
-        suffix = "-" + suffix
+class SaveScreenshotCommand(BaseCommand):
+    def __init__(self, suffix):
+        self.suffix = suffix
 
-    urlhash = md5(driver.current_url.encode("utf-8")).hexdigest()
-    outname = os.path.join(
-        manager_params["screenshot_path"], "%i-%s%s.png" % (visit_id, urlhash, suffix)
-    )
-    driver.save_screenshot(outname)
+    def __repr__(self):
+        return "SaveScreenshotCommand({})".format(self.suffix)
+
+    def execute(
+        self,
+        webdriver,
+        browser_settings,
+        browser_params,
+        manager_params,
+        extension_socket,
+    ):
+        if self.suffix != "":
+            self.suffix = "-" + self.suffix
+
+        urlhash = md5(webdriver.current_url.encode("utf-8")).hexdigest()
+        outname = os.path.join(
+            manager_params["screenshot_path"],
+            "%i-%s%s.png" % (self.visit_id, urlhash, self.suffix),
+        )
+        webdriver.save_screenshot(outname)
 
 
 def _stitch_screenshot_parts(visit_id, browser_id, manager_params):
